fix: resolve compilation errors for Unity 6.5
- Add using UnityEngine.UI to GameManager.cs (CanvasScaler, GraphicRaycaster, etc.)
- Fix AddLegacy -> ApplyLegacy in InterludeManager.cs
- Add Initialize(int) method to UnitManager.cs
- Cast EvaluatePowerRatio to int in DiplomacyManager.cs
- Use int literals in AIPersonalityManager.cs switch pattern
- Add [System.Serializable] to HexCoordinates struct
- Suppress CS0414 unused field warnings in NarrativeScreen.cs

+++ Assets/_Game/Diplomacy/DiplomacyManager.cs
@@ -473,7 +473,7 @@ namespace CivVSCiv
             float allianceRoll = Random.value;
 
             float warThreshold = AIPersonalityManager.GetWarThreshold(personality,
-                EvaluatePowerRatio(aiPlayer, target));
+                (int)EvaluatePowerRatio(aiPlayer, target));
 
             if (warRoll < warThreshold && relations < 0)
             {
